requirements.txt

torch
transformers
pillow
scikit-learn
tensorflow
ultralytics
